Improving coverage

+++ sensorUtil.cxx
@@ -1,15 +1,15 @@
 #include "include/sensorUtil.hxx"
 
-int getAvarage(int * buf, int count) {
+int32_t getAvarage(int32_t * buf, int32_t count) {
     int res = 0;
 
     if (buf == 0 || count == 0) {
         return 0;
     }
 
-    for (int i = 0; i < count; i++) {
+    for (uint16_t i = 0; i < count; i++) {
         res += buf[i];
     }
 
-    return (int) res/count;
+    return (int32_t) res/count;
 }
